Fix issue 57873 Missing translation of the label "Touch to beam"
Root cause:
The SendUi View created when the system first bootup, if we changed
the language settings,the view did not re-create. And the text remained
orignal langguage.
E.G. we set the language to zh-CN, then reboot the phone, active the
SendUi, We see the label show as zh-CN. Then we change the language to
en-US, active the Send View, we find it still show as zh-CN.
Solution:
Set the lable text just before the view displayed.

Change-Id: I004ede4d0109a82f03a6d805062fb35b474e2d3b